🗺️ Atlas: Extract inline tests to tests.rs
